requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
Django==5.2.5
django-cors-headers==4.9.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
google-auth==2.48.0
google-genai==1.66.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.24.0
pillow==11.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
websockets==16.0